+++ Tunny/Solver/Visualize.cs
@@ -20,7 +20,7 @@ namespace Tunny.Solver
             _hasConstraint = hasConstraint;
         }
 
-        private static dynamic LoadStudy(dynamic optuna, string storage, string studyName)
+        private static dynamic LoadStudy(dynamic optuna, dynamic storage, string studyName)
         {
             try
             {
@@ -35,11 +35,11 @@ namespace Tunny.Solver
 
         public void Plot(PlotSettings pSettings)
         {
-            string storage = "sqlite:///" + _settings.Storage.Path;
             PythonEngine.Initialize();
             using (Py.GIL())
             {
                 dynamic optuna = Py.Import("optuna");
+                dynamic storage = _settings.Storage.CreateNewOptunaStorage(false);
                 dynamic study = LoadStudy(optuna, storage, pSettings.TargetStudyName);
                 if (study == null)
                 {
@@ -246,7 +246,7 @@ namespace Tunny.Solver
             return truncate(fig, study);
         }
 
-        private void FigureActions(dynamic fig, PlotSettings pSettings)
+        private static void FigureActions(dynamic fig, PlotSettings pSettings)
         {
             if (fig != null && pSettings.PlotActionType == PlotActionType.Show)
             {
@@ -263,8 +263,8 @@ namespace Tunny.Solver
             var sfd = new SaveFileDialog
             {
                 FileName = name + ".html",
-                Filter = "HTML file(*.html)|*.html",
-                Title = "Save"
+                Filter = @"HTML file(*.html)|*.html",
+                Title = @"Save"
             };
             if (sfd.ShowDialog() == DialogResult.OK)
             {
@@ -274,11 +274,11 @@ namespace Tunny.Solver
 
         public void ClusteringPlot(PlotSettings pSettings)
         {
-            string storage = "sqlite:///" + _settings.Storage.Path;
             PythonEngine.Initialize();
             using (Py.GIL())
             {
                 dynamic optuna = Py.Import("optuna");
+                dynamic storage = _settings.Storage.CreateNewOptunaStorage(false);
                 dynamic study = LoadStudy(optuna, storage, pSettings.TargetStudyName);
                 if (study == null)
                 {